+++ drivers/net/ethernet/mellanox/mlx4/mr.c
@@ -930,7 +930,7 @@ int mlx4_init_mr_table(struct mlx4_dev *dev)
 		return err;
 
 	err = mlx4_buddy_init(&mr_table->mtt_buddy,
-			      ilog2((u32)dev->caps.num_mtts /
+			      ilog2(dev->caps.num_mtts /
 			      (1 << log_mtts_per_seg)));
 	if (err)
 		goto err_buddy;
